+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `debounce` method
 - Add `isUrl` method
+- Add `isObject` method
 
 ## 1.1.1 (2024-11-26)